requirements.txt
brotli==1.1.0
ConfigParser==6.0.0
cryptography==41.0.4
cvzone==1.6.1
Cython==3.0.2
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
importlib_metadata==6.8.0
keyring==24.2.0
lockfile==0.12.2
lxml==4.9.3
monotonic==1.6
numpy==1.26.0
ordereddict==1.1
protobuf==4.24.3
pyOpenSSL==23.2.0
streamlit==1.27.1
tensorflow==2.14.0
mediapipe==0.10.5
Pillow==9.5.0
tempfile2==0.1.2
